Split plugin.py into separate files
Fix #74
